refactor(sip): extract RFC 3261 §17 INVITE client transactions into TransactionLayer
First of five state-machine extractions decomposing the RequestsHandler
monolith. Introduces PbxEnv, the narrow shared-infrastructure interface
(outbox enqueue, message pool, deferred log) the extracted machines talk
back through; RequestsHandler implements it privately. Behavior-neutral:
same slot pool, timers, and call sites — classifyTxType/registerTx are
folded into TransactionLayer::maybeTrack.
