+++ tests/CMakeLists.txt
@@ -79,6 +79,7 @@ add_executable(sip_parser_tests
     ../src/SIP/MixBus.cpp
     ../src/SIP/mix_kernels_scalar.cpp
     ../src/SIP/RequestsHandler.cpp
+    ../src/SIP/TransactionLayer.cpp
     ../src/SIP/SipSdpMessage.cpp
     ../src/SIP/SipClient.cpp
     ../src/SIP/Session.cpp
